test(Gallery): improve unit-tests coverage
