+++ packages/vkui/src/components/BaseGallery/CarouselBase/CarouselBase.test.tsx
@@ -1,13 +1,8 @@
-import * as React from 'react';
-import { fireEvent, render, renderHook } from '@testing-library/react';
-import { noop } from '@vkontakte/vkjs';
-import { AlignType } from '../../../types';
-import { CarouselBase } from './CarouselBase';
+import { renderHook } from '@testing-library/react';
 import { ANIMATION_DURATION } from './constants';
 import { calculateIndent, getLoopPoints, getShiftedIndexes, getTargetIndex } from './helpers';
 import { useSlideAnimation } from './hooks';
 import { SlidesManagerState } from './types';
-import styles from '../BaseGallery.module.css';
 
 const mockRAF = () => {
   let lastTime = 0;
@@ -53,332 +48,6 @@ const createSlidesManager = ({
       })),
   };
 };
-
-const Slide = ({
-  children,
-  width = 200,
-  ...rest
-}: {
-  children: React.ReactNode;
-  width?: number;
-  getRef: React.Ref<HTMLDivElement>;
-  ['data-testid']: string;
-}) => (
-  <div style={{ fontSize: '72px', width }} ref={rest.getRef} data-testid={rest['data-testid']}>
-    {children}
-  </div>
-);
-
-const checkActiveSlide = (container: HTMLElement, slideIndex: number) => {
-  const bullets = Array.from(container.getElementsByClassName(styles['BaseGallery__bullet']));
-  expect(
-    bullets.indexOf(container.getElementsByClassName(styles['BaseGallery__bullet--active'])[0]),
-  ).toBe(slideIndex);
-};
-
-const mockData = ({
-  slideWidth,
-  containerWidth: defaultContainerWidth,
-  viewPortWidth,
-  align,
-  onChange,
-  onNext,
-  onPrev,
-  onDragStart,
-  onDragEnd,
-}: {
-  slideWidth: number;
-  containerWidth: number;
-  viewPortWidth: number;
-  align?: AlignType;
-  onChange: VoidFunction;
-  onNext?: VoidFunction;
-  onPrev?: VoidFunction;
-  onDragStart?: VoidFunction;
-  onDragEnd?: VoidFunction;
-}) => {
-  mockRAF();
-  let slideDataByIndexMap: Record<number, any> = {};
-  let layerTransform = '';
-  let viewPort: HTMLDivElement;
-  let containerWidth = defaultContainerWidth;
-
-  const mockContainerData = (element: HTMLDivElement) => {
-    if (!element) {
-      return;
-    }
-    jest.spyOn(element, 'offsetWidth', 'get').mockImplementation(() => containerWidth);
-  };
-
-  const mockLayerData = (element: HTMLDivElement) => {
-    if (!element) {
-      return;
-    }
-    jest
-      .spyOn(element.style, 'transform', 'set')
-      .mockImplementation((newTransform) => (layerTransform = newTransform));
-  };
-
-  const mockViewportData = (element: HTMLDivElement) => {
-    if (!element) {
-      return;
-    }
-    jest.spyOn(element, 'offsetWidth', 'get').mockReturnValue(viewPortWidth);
-    viewPort = element;
-
-    mockLayerData(element.firstElementChild as HTMLDivElement);
-  };
-
-  const mockSlideData = (element: HTMLElement, index: number) => {
-    if (!element) {
-      return;
-    }
-    jest.spyOn(element.parentElement!, 'offsetWidth', 'get').mockReturnValue(slideWidth);
-    jest.spyOn(element.parentElement!, 'offsetLeft', 'get').mockReturnValue(slideWidth * index);
-
-    let transform = '';
-    jest
-      .spyOn(element.parentElement!.style, 'transform', 'set')
-      .mockImplementation((newTransform) => (transform = newTransform));
-
-    slideDataByIndexMap[index] = {
-      get transform() {
-        return transform;
-      },
-    };
-  };
-
-  const Fixture = ({ slideIndex }: { slideIndex: number }) => (
-    <CarouselBase
-      showArrows
-      align={align}
-      slideIndex={slideIndex}
-      onChange={onChange}
-      onNextClick={onNext}
-      onPrevClick={onPrev}
-      onDragStart={onDragStart}
-      onDragEnd={onDragEnd}
-      bullets="light"
-      getRootRef={mockContainerData}
-      getRef={mockViewportData}
-    >
-      <Slide data-testid="slide-1" getRef={(e: HTMLDivElement) => mockSlideData(e, 0)}>
-        1
-      </Slide>
-      <Slide data-testid="slide-2" getRef={(e: HTMLDivElement) => mockSlideData(e, 1)}>
-        2
-      </Slide>
-      <Slide data-testid="slide-3" getRef={(e: HTMLDivElement) => mockSlideData(e, 2)}>
-        3
-      </Slide>
-      <Slide data-testid="slide-4" getRef={(e: HTMLDivElement) => mockSlideData(e, 3)}>
-        4
-      </Slide>
-      <Slide data-testid="slide-5" getRef={(e: HTMLDivElement) => mockSlideData(e, 4)}>
-        5
-      </Slide>
-    </CarouselBase>
-  );
-
-  const { container, rerender: rerenderFn } = render(<Fixture slideIndex={0} />);
-
-  const rerender = ({ slideIndex }: { slideIndex: number }) => {
-    rerenderFn(<Fixture slideIndex={slideIndex} />);
-  };
-
-  return {
-    container,
-    rerender,
-    get layerTransform() {
-      return layerTransform;
-    },
-    get viewPort() {
-      return viewPort;
-    },
-    getSlideMockData(index: number) {
-      return slideDataByIndexMap[index];
-    },
-    set containerWidth(newWidth: number) {
-      containerWidth = newWidth;
-    },
-  };
-};
-
-describe(CarouselBase, () => {
-  it('check correct navigation by arrows clicks', () => {
-    const onNext = jest.fn();
-    const onPrev = jest.fn();
-    const onChange = jest.fn();
-
-    const mockedData = mockData({
-      slideWidth: 200,
-      containerWidth: 200,
-      viewPortWidth: 200,
-      onPrev,
-      onNext,
-      onChange,
-    });
-    const { container, rerender } = mockedData;
-
-    checkActiveSlide(container, 0);
-
-    const [leftArrow, rightArrow] = Array.from(
-      container.getElementsByClassName(styles['BaseGallery__arrow']),
-    );
-    fireEvent.click(rightArrow);
-
-    expect(onNext).toHaveBeenCalledTimes(1);
-    expect(onChange.mock.calls).toEqual([[1]]);
-
-    rerender({ slideIndex: 1 });
-    checkActiveSlide(container, 1);
-    expect(mockedData.layerTransform).toBe('translate3d(-200px, 0, 0)');
-    fireEvent.click(leftArrow);
-
-    expect(onPrev).toHaveBeenCalledTimes(1);
-    expect(onChange.mock.calls).toEqual([[1], [0]]);
-
-    rerender({ slideIndex: 0 });
-    checkActiveSlide(container, 0);
-    expect(mockedData.layerTransform).toBe('translate3d(0px, 0, 0)');
-
-    fireEvent.click(leftArrow);
-
-    expect(onPrev).toHaveBeenCalledTimes(2);
-    expect(onChange.mock.calls).toEqual([[1], [0], [4]]);
-    rerender({ slideIndex: 4 });
-    checkActiveSlide(container, 4);
-    expect(mockedData.layerTransform).toBe('translate3d(-800px, 0, 0)');
-    expect(mockedData.getSlideMockData(0).transform).toBe('translate3d(0px, 0, 0)');
-  });
-
-  it('check correct navigation by dragging', () => {
-    const onChange = jest.fn();
-    const onDragStart = jest.fn();
-    const onDragEnd = jest.fn();
-
-    const mockedData = mockData({
-      slideWidth: 180,
-      containerWidth: 200,
-      viewPortWidth: 180,
-      align: 'center',
-      onDragStart,
-      onDragEnd,
-      onChange,
-    });
-    const { container, rerender } = mockedData;
-
-    checkActiveSlide(container, 0);
-    fireEvent.mouseDown(mockedData.viewPort, {
-      clientX: 0,
-    });
-    fireEvent.mouseMove(mockedData.viewPort, {
-      clientX: 150,
-    });
-    fireEvent.mouseUp(mockedData.viewPort);
-
-    expect(onDragStart).toHaveBeenCalledTimes(1);
-    expect(onDragEnd).toHaveBeenCalledTimes(1);
-    expect(onChange).toHaveBeenCalledWith(4);
-    rerender({ slideIndex: 4 });
-
-    expect(mockedData.layerTransform).toBe('translate3d(-710px, 0, 0)');
-    expect(mockedData.getSlideMockData(0).transform).toBe('translate3d(900px, 0, 0)');
-    checkActiveSlide(container, 4);
-
-    fireEvent.mouseDown(mockedData.viewPort, {
-      clientX: 150,
-    });
-    fireEvent.mouseMove(mockedData.viewPort, {
-      clientX: 0,
-    });
-    fireEvent.mouseUp(mockedData.viewPort);
-
-    expect(onDragStart).toHaveBeenCalledTimes(2);
-    expect(onDragEnd).toHaveBeenCalledTimes(2);
-    expect(onChange.mock.calls).toEqual([[4], [0]]);
-    rerender({ slideIndex: 0 });
-
-    expect(mockedData.layerTransform).toBe('translate3d(10px, 0, 0)');
-    expect(mockedData.getSlideMockData(0).transform).toBe('translate3d(0px, 0, 0)');
-    checkActiveSlide(container, 0);
-  });
-
-  it('should not change slide when slide too small', () => {
-    const onChange = jest.fn();
-    const onDragStart = jest.fn();
-    const onDragEnd = jest.fn();
-
-    const mockedData = mockData({
-      slideWidth: 180,
-      containerWidth: 200,
-      viewPortWidth: 180,
-      align: 'center',
-      onDragStart,
-      onDragEnd,
-      onChange,
-    });
-    const { container } = mockedData;
-
-    checkActiveSlide(container, 0);
-    fireEvent.mouseDown(mockedData.viewPort, {
-      clientX: 0,
-    });
-    fireEvent.mouseMove(mockedData.viewPort, {
-      clientX: 2,
-    });
-    fireEvent.mouseUp(mockedData.viewPort);
-
-    expect(onDragStart).toHaveBeenCalledTimes(1);
-    expect(onDragEnd).toHaveBeenCalledTimes(1);
-    expect(onChange).toHaveBeenCalledTimes(0);
-
-    expect(mockedData.layerTransform).toBe('translate3d(10px, 0, 0)');
-    expect(mockedData.getSlideMockData(0).transform).toBe('translate3d(0px, 0, 0)');
-  });
-
-  it('should resize when window resize', () => {
-    const onChange = jest.fn();
-
-    const mockedData = mockData({
-      slideWidth: 180,
-      containerWidth: 200,
-      viewPortWidth: 180,
-      align: 'center',
-      onChange,
-    });
-
-    expect(mockedData.layerTransform).toBe('translate3d(10px, 0, 0)');
-    expect(mockedData.getSlideMockData(0).transform).toBe('translate3d(0px, 0, 0)');
-
-    mockedData.containerWidth = 250;
-
-    fireEvent.resize(window);
-
-    expect(mockedData.layerTransform).toBe('translate3d(35px, 0, 0)');
-    expect(mockedData.getSlideMockData(0).transform).toBe('translate3d(0px, 0, 0)');
-  });
-
-  it('check dev error when slides width incorrect', () => {
-    process.env.NODE_ENV = 'development';
-    const onChange = jest.fn();
-    const warn = jest.spyOn(console, 'warn').mockImplementation(noop);
-
-    mockData({
-      slideWidth: 40,
-      containerWidth: 200,
-      viewPortWidth: 200,
-      onChange,
-    });
-
-    expect(warn).toHaveBeenCalledWith(
-      '%c[VKUI/Gallery] Ширины слайдов недостаточно для корректной работы свойства "looped". Пожалуйста, сделайте её больше."',
-      undefined,
-    );
-    warn.mockRestore();
-    process.env.NODE_ENV = 'test';
-  });
-});
 
 describe(calculateIndent, () => {
   it.each([
